requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
slack-sdk>=3.27.0
python-dotenv>=1.0.0
pytest>=8.0.0
